Refactor: 연간리포트 집계함수 사용하여 최적화

+++ src/main/java/com/project/wallet_keeper/dto/transaction/AnnualSummary.java
@@ -15,8 +15,8 @@ import java.util.Map;
 public class AnnualSummary implements Serializable {
 
     private int year;
-    private int totalIncome;
-    private int totalExpense;
-    private int total;
+    private long totalIncome;
+    private long totalExpense;
+    private long total;
     private Map<String, MonthlySummary> monthly;
 }
